requirements.txt
aiohttp==3.6.2
Django==3.0.6
djangorestframework==3.11.0
drf-yasg==1.17.1
factory-boy==2.12.0
psycopg2==2.8.5
requests==2.23.0
